pyproject.toml
[project]
name = "study-helper"
version = "0.1.0"
description = "숭실대학교 LMS 강의 백그라운드 재생 및 다운로드/요약 도구"
requires-python = ">=3.11,<3.13"

dependencies = [
    # 브라우저 자동화 (로그인, 스크래핑, 백그라운드 재생)
    "playwright>=1.52.0",
    # CUI 인터페이스
    "rich>=13.0.0",
    "click>=8.0.0",
    # 영상/음성 다운로드 (청크 스트리밍)
    "requests>=2.32.0",
    # STT — Whisper에 mp4 직접 전달 (wav 중간 파일 불필요)
    # torch는 Dockerfile에서 CPU 전용으로 별도 설치
    "openai-whisper>=20240930",
    # AI 요약
    "google-generativeai>=0.8.0",
    "openai>=1.0.0",
    # 환경변수
    "python-dotenv>=1.0.0",
    "cryptography>=46.0.5",
]

# torch는 Dockerfile에서 CPU wheel로 직접 설치하므로 여기서 제외
# (Docker 외 로컬 실행 시에는 별도로 pip install torch 필요)
[project.optional-dependencies]
local = [
    "torch>=2.0.0",
]

[project.scripts]
study-helper = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
